+++ configuration/src/agent/signer.rs
@@ -3,7 +3,7 @@ use std::str::FromStr;
 use crate::FromEnv;
 use nomad_types::HexString;
 
-/// Signer types
+/// Ethereum signer types
 #[derive(Debug, Clone, PartialEq, serde::Deserialize)]
 #[serde(untagged, rename_all = "camelCase")]
 pub enum SignerConf {
@@ -76,5 +76,29 @@ mod test {
                 region: "".to_owned()
             }
         );
+    }
+}
+
+impl SignerConf {
+    /// Validate signer conf fields
+    pub fn validate(&self, network: &str) -> eyre::Result<()> {
+        Ok(match self {
+            SignerConf::HexKey { key } => {
+                eyre::ensure!(
+                    !key.as_ref().is_empty(),
+                    "Hex signer key for {} empty!",
+                    network,
+                );
+            }
+            SignerConf::Aws { id, region } => {
+                eyre::ensure!(!id.is_empty(), "ID for {} aws signer key empty!", network,);
+                eyre::ensure!(
+                    !region.is_empty(),
+                    "Region for {} aws signer key empty!",
+                    network,
+                );
+            }
+            SignerConf::Node => (),
+        })
     }
 }